requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
flask==3.0.0
flask-session==0.5.0
flask-limiter==3.5.0

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# AI & Data Processing
google-generativeai==0.3.1
pandas==2.1.3
numpy==1.25.2

# Rate Limiting & Concurrency
ratelimit==2.2.1
asyncio-throttle==1.0.2

# Production Dependencies
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4

# Monitoring & Observability
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Security
python-multipart==0.0.6
cryptography==41.0.7

# Database (Optional)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Caching
aioredis==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1